Trying to make DU data nodes work. Summary:
Approach is with a parallel tag union type as cache key for the node
data types.

This feels like a dead-end. I'm going to struggle to relate the parallel tag
union-type to the union-type itself in the cache. Again, the problem is that I
cannot talk about the union cases separately - union type instances are either
the union type or not, they can be unwrapped but otherwise don't stand on their
own as a type. I feel like this is not a good way to implement the idea of a
'sum-type' as it's not a sum of types, but rather a collection of formless
(typeless) entities which do not stand separately. A sum of things is a
collection of things, ie the things themselves can be talked about. Here it's
impossible to refer to individual cases relative to their existence as part of
the collection. There is a fslang suggestion about making case types available
but it's stagnated. https://github.com/fsharp/fslang-suggestions/issues/606